requirements.txt
fastapi[all]
passlib[bcrypt]
python-jose[cryptography]
alembic
email-validator
psycopg2-binary